Move service creation from install to configure
Having the service creation in the install tasks means that changes to
the configuration, such as the web.external-url, are not updated with
the configuration task.  This creates a problem if configuration is
re-applied, either as a separate step, or re-run for updating values.
Moving the service creation to the configure step ensures any new or
updated values are taken into account in the service execution.